Embed effect system as new damage model
