+++ test/lib/effect.spec.js
@@ -1,135 +1,5 @@
 import { expect } from "chai";
-
-const distance = (x1, y1, x2, y2) =>
-  Math.sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
-
-const normalize = (x1, y1, x2, y2) => {
-  const dist = distance(x1, y1, x2, y2);
-  return [(x2 - x1) / dist, (y2 - y1) / dist];
-};
-
-const addEffect = (
-  target,
-  {
-    affects,
-    duration = 1,
-    velocity = 0,
-    accelleration = 0,
-    upperBounds = Infinity,
-    lowerBounds = -Infinity,
-    origin,
-    speed,
-    name = "Impact"
-  }
-) => ({
-  ...target,
-  effects: (target.effects || []).concat({
-    affects,
-    duration,
-    velocity,
-    accelleration,
-    upperBounds,
-    lowerBounds,
-    active: 0,
-    origin,
-    speed,
-    name
-  })
-});
-
-const processEffects = (worldRules = {}) => (target, duration) => {
-  const { mutations, effects } = target.effects.reduce(
-    (acc, effect) => {
-      let travelDuration = Infinity;
-
-      if (
-        effect.origin &&
-        effect.speed &&
-        target.x !== undefined &&
-        target.y !== undefined
-      ) {
-        const dist = distance(
-          target.x,
-          target.y,
-          effect.origin.x,
-          effect.origin.y
-        );
-        const traveled = ((effect.active + duration) / 1000) * effect.speed;
-        if (dist < traveled) {
-          travelDuration = ((traveled - dist) / effect.speed) * 1000;
-        } else {
-          travelDuration = 0;
-        }
-      }
-
-      const calcDuration =
-        Math.min(duration, travelDuration, effect.duration) / 1000;
-
-      const worldEffect = (prop, amount) => {
-        if (worldRules[prop]) {
-          return worldRules[prop](amount, effect, target);
-        }
-        return { [prop]: amount };
-      };
-
-      const delta =
-        effect.velocity * calcDuration +
-        0.5 * effect.accelleration * calcDuration * calcDuration;
-
-      const clippedDelta =
-        delta > effect.upperBounds
-          ? effect.upperBounds
-          : delta < effect.lowerBounds
-          ? effect.lowerBounds
-          : delta;
-
-      const mutations =
-        clippedDelta !== 0
-          ? acc.mutations.concat(
-              []
-                .concat(effect.affects)
-                .reduce(
-                  (acc, prop) =>
-                    acc.concat(
-                      Object.entries(worldEffect(prop, clippedDelta)).reduce(
-                        (acc, [prop, amount]) =>
-                          acc.concat([[prop, e => e + amount]]),
-                        []
-                      )
-                    ),
-                  []
-                )
-            )
-          : acc.mutations;
-
-      return acc.mutations !== mutations ? { ...acc, mutations } : acc;
-    },
-    {
-      mutations: [],
-      effects: target.effects.reduce(
-        (acc, effect) =>
-          effect.duration > duration
-            ? acc.concat({
-                ...effect,
-                duration: effect.duration - duration,
-                active: effect.active + duration,
-                velocity:
-                  effect.velocity + (effect.accelleration * duration) / 1000
-              })
-            : acc,
-        []
-      )
-    }
-  );
-
-  return mutations.reduce(
-    (acc, [prop, f]) => {
-      acc[prop] = f(acc[prop]);
-      return acc;
-    },
-    { ...target, effects }
-  );
-};
+import { normalize, addEffect, processEffects } from "../../src/lib/effects";
 
 /**
  * Goals:
@@ -317,11 +187,14 @@ describe("Effect model", () => {
     expect(result.life).to.eq(72); // instead of 60
     expect(result.mana).to.eq(360);
 
-    const result2 = applyEffects(addEffect(armoredTarget, damage), 1);
+    const result2 = applyEffects(addEffect({ ...armoredTarget }, damage), 1);
     expect(result2.life).to.eq(96); // instead of 60
     expect(result2.mana).to.eq(360);
 
-    const poisonResult = applyEffects(addEffect(armoredTarget, poison), 1);
+    const poisonResult = applyEffects(
+      addEffect({ ...armoredTarget }, poison),
+      1
+    );
     expect(poisonResult.life).to.eq(60); // no resistance
     expect(poisonResult.mana).to.eq(360);
   });
@@ -347,10 +220,12 @@ describe("Effect model", () => {
         name: "Blast"
       };
 
-      const result = applyEffects(addEffect(target, damage), 10);
+      const changes = applyEffects(addEffect(target, damage), 10);
+      const result = { ...target, ...changes };
       expect(result.life).to.eq(100);
 
-      const result2 = applyEffects(result, 1000);
+      const changes2 = applyEffects(result, 1000);
+      const result2 = { ...result, ...changes2 };
       expect(round2(result2.life)).to.eq(40.67);
       expect(round2(result2.x)).to.eq(10);
     });
@@ -375,10 +250,12 @@ describe("Effect model", () => {
         name: "Blast"
       };
 
-      const result = applyEffects(addEffect(target, damage), 10);
+      const changes = applyEffects(addEffect(target, damage), 10);
+      const result = { ...target, ...changes };
       expect(result.life).to.eq(100);
 
-      const result2 = applyEffects(result, 1000);
+      const changes2 = applyEffects(result, 1000);
+      const result2 = { ...result, ...changes2 };
       expect(round2(result2.life)).to.eq(24.09);
       expect(round2(result2.x)).to.eq(80);
     });
@@ -446,5 +323,24 @@ describe("Effect model", () => {
       expect(round2(result3.x)).to.eq(166.99); // Δ 166.99 - 120 = 46.99
       expect(round2(result3.y)).to.eq(166.99);
     });
+  });
+
+  it("returns only the mutated properties", () => {
+    const applyEffects = processEffects();
+    const target = {
+      life: 100,
+      foo: 842,
+      bar: 1337
+    };
+    const damage = {
+      velocity: -40e3, // DPS, active for 1 ms
+      affects: "life"
+    };
+
+    const result = applyEffects(addEffect(target, damage), 1);
+    expect(result.life).to.eq(60);
+    expect(result.foo).to.be.undefined;
+    expect(result.bar).to.be.undefined;
+    expect(result.effects).to.eql([]);
   });
 });
